Patch multiple processing and data issues
This patch fixes processing erroring, gamp parsing speed, cache loading,
cache overrides, binning, likelihoods, and pipes.

The processing module needed a few design tweaks to properly handle
pipes in a way that wouldn't result in a OSError. Now, the stop method
has been removed from the ProcessInterface, and has been replaced with
a close method that properly shuts down the processes, as well as close
the effected pipes. SmartProcess has been updated as well, so that when
the process closes, or the process encounters an exception, the pipe for
that process will be closed as well. When close is properly called on
ProcessInterface after the objects are done being used, it shouldn't be
possible to encounter the OSError anymore.

Processing also supported sending a ProcessCode when an error was
encountered, but the interfaces wouldn't search for the ProcessCode, and
would instead throw an error when it encountered it, suppressing the real
error. Now, the interfaces correctly search for an handle the error, as
well as forwarding the exception so that the user can patch their
amplitude.

Caching has been patched as well. You can now forcefully disable the
cache, or overwrite the cache from read and write. The issue with the
caching module resulted from the cache package would overwrite the
created cache package when it was loaded, which meant that when hashes
were compared, they'd always result in the hashes appearing to match.

Gamp memory parser was updated to allow for quicker parsing of Gamp
files. DataFrames are used internally to store the values, but changing
a DataFrames values after it's been created is a relatively time
consuming process. It's significantly faster to create a numpy array and
update its values with what is in the file, and then convert that numpy
array into a DataFrame. This is effectly what Gamp does now, and is
similar to the method used by the Kv parser.

Binning had an issue where it return N bins - 1, now, it will return the
correct number of bins.

The Extended Log Likelihood would return the wrong result due to a
miscalculation in the likelihood. This has been corrected.

+++ PyPWA/libs/process.py
@@ -43,6 +43,7 @@ File Layout
 """
 
 import copy
+import time
 from abc import ABC, abstractmethod
 from enum import Enum
 from multiprocessing import cpu_count, Pipe, Process
@@ -220,34 +221,35 @@ class ProcessCodes(Enum):
 
 class ProcessInterface:
 
-    def __init__(self,
-                 interface_kernel: Interface,
-                 process_com: List[Connection],
-                 processes: List["_SmartProcess"]):
+    def __init__(
+            self, interface_kernel: Interface,
+            process_com: List[Connection], processes: List["_SmartProcess"]):
         self.__connections = process_com
         self.__interface = interface_kernel
         self.__processes = processes
 
     def run(self, *args):
-        return self.__interface.run(self.__connections, args)
+        try:
+            return self.__interface.run(self.__connections, args)
+        except Exception as error:
+            self.close()
+            raise error
 
-    def stop(self, force: bool = False):
-        if self.__connections[0].writable and not force:
-            self.__ask_processes_to_stop()
-        else:
-            self.__terminate_processes()
-
-    def __ask_processes_to_stop(self):
+    def close(self):
         for connection in self.__connections:
-            connection.send(ProcessCodes.SHUTDOWN)
+            if connection.writable:
+                connection.send(ProcessCodes.SHUTDOWN)
+            connection.close()
 
-    def __terminate_processes(self):
+        time.sleep(2)
+
         for process in self.__processes:
-            process.terminate()
+            if process.is_alive():
+                process.terminate()
 
     @property
     def is_alive(self) -> bool:
-        return True in [proc.is_alive() for proc in self.__processes]
+        return any([proc.is_alive() for proc in self.__processes])
 
 
 class _SmartProcess(Process):
@@ -267,8 +269,8 @@ class _SmartProcess(Process):
     def __run_duplex(self):
         try:
             self.__kernel.setup()
-        except Exception:
-            self.__connection.send(ProcessCodes.ERROR)
+        except Exception as error:
+            self.__handle_error(error)
             raise
         else:
             self.__loop()
@@ -277,14 +279,15 @@ class _SmartProcess(Process):
         while True:
             received = self.__connection.recv()
             if received == ProcessCodes.SHUTDOWN:
+                self.__connection.close()
                 break
             self.__process(received)
 
     def __process(self, received_data):
         try:
             value = self.__kernel.process(received_data)
-        except Exception:
-            self.__connection.send(ProcessCodes.ERROR)
+        except Exception as error:
+            self.__handle_error(error)
             raise
         else:
             self.__connection.send(value)
@@ -293,6 +296,11 @@ class _SmartProcess(Process):
         try:
             self.__kernel.setup()
             self.__connection.send(self.__kernel.process())
-        except Exception:
-            self.__connection.send(ProcessCodes.ERROR)
+        except Exception as error:
+            self.__handle_error(error)
             raise
+
+    def __handle_error(self, error):
+        self.__connection.send(ProcessCodes.ERROR)
+        self.__connection.send(error)
+        self.__connection.close()
